+++ CHANGELOG.md
@@ -214,3 +214,4 @@ Fix of `CalculatedDateTime#asPastFor` - now it will decrease time even if source
 * Now rating receivers will automatically register message for post as likes message if have no registered
 * `TimerTriggerStrategy` now do not using synchronisations
 * Add `Iterable<CalculatedDateTime>#nearDateTime`
+* `TimerTriggerStrategy` config now receive time as parameter